requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
tqdm>=4.60.0
astropy>=5.0
pytest>=7.0.0
jupyter>=1.0.0
